finsight_ai/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coral-backend"
version = "0.1.0"
description = "Coral - Local-first financial intelligence backend"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Data validation & settings
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",

    # Database
    "sqlmodel>=0.0.21",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",

    # PDF parsing
    "pdfplumber>=0.11.0",
    "pypdf>=4.3.0",

    # Vector database
    "chromadb>=0.5.0",

    # LangGraph + LangChain
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",

    # Ollama client
    "ollama>=0.3.0",

    # Async HTTP
    "httpx>=0.27.0",

    # Utilities
    "python-dateutil>=2.9.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "rich>=13.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B"]
ignore = ["B008"]  # FastAPI Depends() in function args

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
